refactoring: split existing file into two more files
