+++ packages/typir/src/kinds/function/function-initializer.ts
@@ -6,15 +6,14 @@
 
 import { isType, Type, TypeStateListener } from '../../graph/type-node.js';
 import { TypeInitializer } from '../../initialization/type-initializer.js';
-import { AssignabilitySuccess, isAssignabilityProblem } from '../../services/assignability.js';
-import { isConversionEdge } from '../../services/conversion.js';
-import { CompositeTypeInferenceRule, InferenceProblem, InferenceRuleNotApplicable, TypeInferenceRule, TypeInferenceRuleWithInferringChildren } from '../../services/inference.js';
-import { isSubTypeEdge } from '../../services/subtype.js';
+import { CompositeTypeInferenceRule, InferenceProblem, InferenceRuleNotApplicable, TypeInferenceRule } from '../../services/inference.js';
 import { ValidationRule } from '../../services/validation.js';
 import { TypirServices } from '../../typir.js';
 import { checkTypeArrays } from '../../utils/utils-type-comparison.js';
-import { assertType, assertUnreachable } from '../../utils/utils.js';
+import { assertType } from '../../utils/utils.js';
+import { FunctionCallInferenceRule } from './function-call-inference.js';
 import { CreateFunctionTypeDetails, FunctionKind, OverloadedFunctionDetails } from './function-kind.js';
+import { OverloadedFunctionsTypeInferenceRule } from './function-overloaded-inference.js';
 import { FunctionType, isFunctionType } from './function-type.js';
 
 export class FunctionTypeInitializer<T> extends TypeInitializer<FunctionType> implements TypeStateListener {
@@ -209,218 +208,4 @@ interface FunctionInferenceRules {
     inferenceForCall?: TypeInferenceRule;
     validationForCall?: ValidationRule;
     inferenceForDeclaration?: TypeInferenceRule;
-}
-
-
-/**
- * Dedicated inference rule for calls of a single function signature.
- * It takes into account, that all parameters match and provides information, how parameters are matching ('assignabilitySuccess').
- *
- * Preconditions:
- * - there is a rule which specifies how to infer the current function type
- * - the current function has an output type/parameter, otherwise, this function could not provide any type (and throws an error), when it is called!
- *   (exception: the options contain a type to return in this special case)
- */
-class FunctionCallInferenceRule<T> implements TypeInferenceRuleWithInferringChildren {
-    protected readonly typeDetails: CreateFunctionTypeDetails<T>;
-    protected readonly functionType: FunctionType;
-    protected readonly mapNameTypes: Map<string, OverloadedFunctionDetails>;
-    assignabilitySuccess: Array<AssignabilitySuccess | undefined>;
-
-    constructor(typeDetails: CreateFunctionTypeDetails<T>, functionType: FunctionType, mapNameTypes: Map<string, OverloadedFunctionDetails>) {
-        this.typeDetails = typeDetails;
-        this.functionType = functionType;
-        this.mapNameTypes = mapNameTypes;
-        this.assignabilitySuccess = new Array(typeDetails.inputParameters.length);
-    }
-
-    inferTypeWithoutChildren(languageNode: unknown, _typir: TypirServices): unknown {
-        this.assignabilitySuccess.fill(undefined); // reset the entries
-        // 1. Does the filter of the inference rule accept the current language node?
-        const result = this.typeDetails.inferenceRuleForCalls!.filter(languageNode);
-        if (!result) {
-            // the language node has a completely different purpose
-            return InferenceRuleNotApplicable;
-        }
-        // 2. Does the inference rule match this language node?
-        const matching = this.typeDetails.inferenceRuleForCalls!.matching(languageNode);
-        if (!matching) {
-            // the language node is slightly different
-            return InferenceRuleNotApplicable;
-        }
-        // 3. Check whether the current arguments fit to the expected parameter types
-        const inputArguments = this.typeDetails.inferenceRuleForCalls!.inputArguments(languageNode);
-        if (inputArguments.length <= 0) {
-            // there are no operands to check
-            return this.check(this.getOutputTypeForFunctionCalls());
-        }
-        // at least one operand => this function type might match, to be sure, resolve the types of the values for the parameters
-        const overloadInfos = this.mapNameTypes.get(this.typeDetails.functionName);
-        if (overloadInfos === undefined || overloadInfos.overloadedFunctions.length <= 1) {
-            // the current function is not overloaded, therefore, the types of their parameters are not required => save time, ignore inference errors
-            return this.check(this.getOutputTypeForFunctionCalls());
-        }
-        // two or more overloaded functions
-        if (overloadInfos.sameOutputType) {
-            // exception: all(!) overloaded functions have the same(!) output type, save performance and return this type!
-            return overloadInfos.sameOutputType;
-        }
-        // the types of the parameters need to be inferred in order to determine an exact match
-        return inputArguments;
-    }
-
-    inferTypeWithChildrensTypes(languageNode: unknown, actualInputTypes: Array<Type | undefined>, typir: TypirServices): Type | InferenceProblem {
-        const expectedInputTypes = this.typeDetails.inputParameters.map(p => typir.infrastructure.TypeResolver.resolve(p.type));
-        // all operands need to be assignable(! not equal) to the required types
-        const comparisonConflicts = checkTypeArrays(
-            actualInputTypes,
-            expectedInputTypes,
-            (t1, t2, index) => {
-                const result = typir.Assignability.getAssignabilityResult(t1, t2);
-                if (isAssignabilityProblem(result)) {
-                    return result;
-                } else {
-                    // save the information equal/conversion/subtype for deciding "conflicts" of overloaded functions
-                    this.assignabilitySuccess[index] = result;
-                    return undefined;
-                }
-            },
-            true,
-        );
-        if (comparisonConflicts.length >= 1) {
-            // this function type does not match, due to assignability conflicts => return them as errors
-            return {
-                $problem: InferenceProblem,
-                languageNode: languageNode,
-                inferenceCandidate: this.functionType,
-                location: 'input parameters',
-                rule: this,
-                subProblems: comparisonConflicts,
-            };
-            // We have a dedicated validation for this case (see below), but a resulting error might be ignored by the user => return the problem during type-inference again
-        } else {
-            // matching => return the return type of the function for the case of a function call!
-            return this.check(this.getOutputTypeForFunctionCalls());
-        }
-    }
-
-    protected getOutputTypeForFunctionCalls(): Type | undefined {
-        return this.functionType.kind.getOutputTypeForFunctionCalls(this.functionType);
-    }
-
-    protected check(returnType: Type | undefined): Type {
-        if (returnType) { // this condition is checked here, since 'undefined' is OK, as long as it is not used; extracting this function is difficult due to TypeScripts strict rules for using 'this'
-            return returnType;
-        } else {
-            throw new Error(`The function ${this.typeDetails.functionName} is called, but has no output type to infer.`);
-        }
-    }
-}
-
-
-/**
- * Custom inference rule for functions, which consists of one inference rule for each overload/signature for a function with same name.
- * When deadling with multiple inference rules, usually the first successful inference rule is applied and following inference rules are ignored.
- * In order to deal with multiple matching inference rules for overloaded functions,
- * all available inference rules need to be executed and all successful inference rules need to be collected.
- */
-export class OverloadedFunctionsTypeInferenceRule extends CompositeTypeInferenceRule {
-
-    protected override inferTypeLogic(languageNode: unknown): Type | InferenceProblem[] {
-        this.checkForError(languageNode);
-
-        // check all rules in order to search for the best-matching rule, not for the first-matching rule
-        const matchingOverloads: OverloadedMatch[] = [];
-        const collectedInferenceProblems: InferenceProblem[] = [];
-        for (const rules of this.inferenceRules.values()) {
-            for (const rule of rules) {
-                const result = this.executeSingleInferenceRuleLogic(rule, languageNode, collectedInferenceProblems);
-                if (result) {
-                    matchingOverloads.push({ result, rule: rule as FunctionCallInferenceRule<never> });
-                } else {
-                    // no result for this inference rule => check the next inference rules
-                }
-            }
-        }
-
-        if (matchingOverloads.length <= 0) {
-            // no matches => return all the collected inference problems
-            if (collectedInferenceProblems.length <= 0) {
-                // document the reason, why neither a type nor inference problems are found
-                collectedInferenceProblems.push({
-                    $problem: InferenceProblem,
-                    languageNode: languageNode,
-                    location: 'found no applicable inference rules',
-                    subProblems: [],
-                });
-            }
-            return collectedInferenceProblems;
-        } else if (matchingOverloads.length === 1) {
-            // single match
-            return matchingOverloads[0].result;
-        } else {
-            // multiple matches => determine the one to return
-
-            // 1. identify and collect the best matches
-            const bestMatches: OverloadedMatch[] = [ matchingOverloads[0] ];
-            for (let i = 1; i < matchingOverloads.length; i++) {
-                const currentMatch = matchingOverloads[i];
-                const comparison = this.compareMatchingOverloads(bestMatches[0], currentMatch);
-                if (comparison < 0) {
-                    // the existing matches are better than the current one => keep the existing best matches
-                } else if (comparison > 0) {
-                    // the current match is better than the already collect ones => replace the existing best matches by the current one
-                    bestMatches.splice(0, bestMatches.length, currentMatch);
-                } else {
-                    // the current and the existing matches are both good => collect both
-                    bestMatches.push(currentMatch);
-                }
-            }
-
-            // 2. evaluate the remaining best matches
-            if (bestMatches.length === 0) {
-                // return the single remaining match
-                return bestMatches[0].result;
-            } else {
-                // decide how to deal with multiple best matches
-                const result = this.handleMultipleBestMatches(bestMatches);
-                if (result) {
-                    // return the chosen match
-                    return result.result;
-                } else {
-                    // no decision => inference is not possible
-                    return [{
-                        $problem: InferenceProblem,
-                        languageNode: languageNode,
-                        location: `Found ${bestMatches.length} best matching overloads: ${bestMatches.map(m => m.result.getIdentifier()).join(', ')}`,
-                        subProblems: [], // there are no real sub-problems, since the relevant overloads match ...
-                    }];
-                }
-            }
-        }
-    }
-
-    protected handleMultipleBestMatches(matchingOverloads: OverloadedMatch[]): OverloadedMatch | undefined {
-        return matchingOverloads[0]; // by default, return the 1st best match
-    }
-
-    // better matches are at the beginning of the list, i.e. better matches get values lower than zero
-    protected compareMatchingOverloads(match1: OverloadedMatch, match2: OverloadedMatch): number {
-        const cost1 = this.calculateCost(match1);
-        const cost2 = this.calculateCost(match2);
-        return cost1 === cost2 ? 0 : cost1 < cost2 ? -1 : +1;
-    }
-
-    protected calculateCost(match: OverloadedMatch): number {
-        return match.rule.assignabilitySuccess
-            .flatMap(s => s?.path ?? []) // collect all conversion/sub-type edges which are required to map actual types to the expected types of the parameters
-            // equal types (i.e. an empty path) are better than sub-types, sub-types are better than conversions
-            .map(edge => (isSubTypeEdge(edge) ? 1 : isConversionEdge(edge) ? 2 : assertUnreachable(edge)) as number)
-            .reduce((l, r) => l + r, 0);
-    }
-}
-
-interface OverloadedMatch {
-    result: Type;
-    rule: FunctionCallInferenceRule<never>;
 }
